% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_optimise.R
\name{suggested_breakpoints_vector}
\alias{suggested_breakpoints_vector}
\title{Extract the selected breakpoints from suggest_breakpoints() as a vector}
\usage{
suggested_breakpoints_vector(candidates_dt)
}
\arguments{
\item{candidates_dt}{The data.table returned by \code{\link[=suggest_breakpoints]{suggest_breakpoints()}}.}
}
\value{
A numeric vector, sorted ascending.
}
\description{
Extract the selected breakpoints from suggest_breakpoints() as a vector
}
